Move setup tasks to their own file
